Write access pattern functions for 1D, 2D, 3D slices
These functions should be combined with patch
indexing functions to enable traversal of patch
data in the index (not memory) space. The simplest
way to do this is to use i,j,k indexing on the
arrays, so that method should be done first.
However access patterns built from iterators avoid
unnecessary summation of the linear index, so they
could turn out to be faster.