Add a solvers sub-module: surprise bump to 17.0 Mzps
It must be because when the solver was in the example
code, it was not optimized together with the hydro
module in the library. Something to keep in mind:
the solver and hydro modules should be in the same
crate, or otherwise either the hydro functions should
all be inlined, or LTO needs to be enabled.